Added tests for cli.
